+++ extend/waibuBootstrap/theme/component/widget/control-logo.js
@@ -29,7 +29,7 @@ async function controlLogo () {
       }
       this.addBlock('reactive', `
         async ${prefix}Builder () {
-          const body = ['<c:a href="#" @click="wbs.appLauncher(\\'darkmode language\\', \\'${this.params.attr.menu}\\', { theme: \\'${this.component.theme.name}\\', iconset: \\'${this.component.iconset.name}\\' })">']
+          const body = ['<c:a href="#" @click="wbs.appLauncher(\\'darkmode language\\', \\'${this.params.attr.menu}\\')">']
           body.push('<c:img ${stringifyAttribs(img)} ${animate} />')
           body.push('</c:a>')
           return await wmpa.createComponent(body)